// src/generated/api/models/UserProfileResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface UserProfileResponse
 */
export interface UserProfileResponse {
  /**
   * Github 아이디
   * @type {string}
   * @memberof UserProfileResponse
   */
  githubId?: string;
  /**
   * 닉네임
   * @type {string}
   * @memberof UserProfileResponse
   */
  nickname?: string;
  /**
   * 이름
   * @type {string}
   * @memberof UserProfileResponse
   */
  name?: string;
  /**
   * 이메일
   * @type {string}
   * @memberof UserProfileResponse
   */
  email?: string;
  /**
   * 프로필 이미지 URL
   * @type {string}
   * @memberof UserProfileResponse
   */
  profileImageUrl?: string;
  /**
   * 직책/포지션
   * @type {string}
   * @memberof UserProfileResponse
   */
  position?: string;
  /**
   * 유저 역할
   * @type {string}
   * @memberof UserProfileResponse
   */
  role?: string;
  /**
   * 학번
   * @type {string}
   * @memberof UserProfileResponse
   */
  studentNumber?: string;
  /**
   * 감자 동아리 기수
   * @type {number}
   * @memberof UserProfileResponse
   */
  gamjaBatch?: number;
  /**
   * 생성 일시
   * @type {string}
   * @memberof UserProfileResponse
   */
  createdAt?: string;
  /**
   * 수정 일시
   * @type {string}
   * @memberof UserProfileResponse
   */
  updatedAt?: string;
}

/**
 * Check if a given object implements the UserProfileResponse interface.
 */
export function instanceOfUserProfileResponse(value: object): value is UserProfileResponse {
  return true;
}

export function UserProfileResponseFromJSON(json: any): UserProfileResponse {
  return UserProfileResponseFromJSONTyped(json, false);
}

export function UserProfileResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): UserProfileResponse {
  if (json == null) {
    return json;
  }
  return {
    githubId: json["githubId"] == null ? undefined : json["githubId"],
    nickname: json["nickname"] == null ? undefined : json["nickname"],
    name: json["name"] == null ? undefined : json["name"],
    email: json["email"] == null ? undefined : json["email"],
    profileImageUrl: json["profileImageUrl"] == null ? undefined : json["profileImageUrl"],
    position: json["position"] == null ? undefined : json["position"],
    role: json["role"] == null ? undefined : json["role"],
    studentNumber: json["studentNumber"] == null ? undefined : json["studentNumber"],
    gamjaBatch: json["gamjaBatch"] == null ? undefined : json["gamjaBatch"],
    createdAt: json["createdAt"] == null ? undefined : json["createdAt"],
    updatedAt: json["updatedAt"] == null ? undefined : json["updatedAt"],
  };
}

export function UserProfileResponseToJSON(json: any): UserProfileResponse {
  return UserProfileResponseToJSONTyped(json, false);
}

export function UserProfileResponseToJSONTyped(
  value?: UserProfileResponse | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    githubId: value["githubId"],
    nickname: value["nickname"],
    name: value["name"],
    email: value["email"],
    profileImageUrl: value["profileImageUrl"],
    position: value["position"],
    role: value["role"],
    studentNumber: value["studentNumber"],
    gamjaBatch: value["gamjaBatch"],
    createdAt: value["createdAt"],
    updatedAt: value["updatedAt"],
  };
}