// src/generated/api/models/ResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 모든 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0.4
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
import type { ProjectListResponse } from "./ProjectListResponse";
import {
  ProjectListResponseFromJSON,
  ProjectListResponseFromJSONTyped,
  ProjectListResponseToJSON,
  ProjectListResponseToJSONTyped,
} from "./ProjectListResponse";

/**
 * 공통 API 응답 래퍼
 * @export
 * @interface ResponseDto
 */
export interface ResponseDto {
  /**
   *
   * @type {number}
   * @memberof ResponseDto
   */
  status?: number;
  /**
   *
   * @type {string}
   * @memberof ResponseDto
   */
  message?: string;
  /**
   *
   * @type {ProjectListResponse}
   * @memberof ResponseDto
   */
  data?: ProjectListResponse;
  /**
   *
   * @type {number}
   * @memberof ResponseDto
   */
  timestamp?: number;
}

/**
 * Check if a given object implements the ResponseDto interface.
 */
export function instanceOfResponseDto(value: object): value is ResponseDto {
  return true;
}

export function ResponseDtoFromJSON(json: any): ResponseDto {
  return ResponseDtoFromJSONTyped(json, false);
}

export function ResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResponseDto {
  if (json == null) {
    return json;
  }
  return {
    status: json["status"] == null ? undefined : json["status"],
    message: json["message"] == null ? undefined : json["message"],
    data: json["data"] == null ? undefined : ProjectListResponseFromJSON(json["data"]),
    timestamp: json["timestamp"] == null ? undefined : json["timestamp"],
  };
}

export function ResponseDtoToJSON(json: any): ResponseDto {
  return ResponseDtoToJSONTyped(json, false);
}

export function ResponseDtoToJSONTyped(
  value?: ResponseDto | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    status: value["status"],
    message: value["message"],
    data: ProjectListResponseToJSON(value["data"]),
    timestamp: value["timestamp"],
  };
}